fix: move SSL params classes to standalone header to resolve circular includes
Servers_SslParams, MySQLServers_SslParams, and PgSQLServers_SslParams
are now in Servers_SslParams.h with minimal dependencies (<string> only).
This avoids the circular include issue between cpp.h, mysql_connection.h,
and Base_HostGroups_Manager.h that caused incomplete type errors.
